requirements.txt
streamlit==1.20.0
tensorflow==2.6.0
Pillow==8.3.2